Fix review comments
