--- docs/reference/sql/st_applydefaultcrs.qmd
@@ -1,43 +1,0 @@
----
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-title: ST_ApplyDefaultCRS
-description: Stamps the default CRS (OGC:CRS84) onto a geometry that does not already have one.
-kernels:
-  - returns: geometry
-    args:
-    - geometry
----
-
-## Description
-
-Sets the default CRS (OGC:CRS84, equivalent to EPSG:4326) on a geometry that does not already have one. If the input geometry already carries a CRS, it is passed through unchanged.
-
-This function is useful for assigning a CRS to geometries that were created without one, so that they can be used with CRS-aware functions such as RS_Intersects, RS_Contains, and RS_Within.
-
-## Examples
-
-```sql
--- Assign the default CRS to a point that has no CRS
-SELECT ST_ApplyDefaultCRS(ST_Point(-122.35, 47.65));
-```
-
-```sql
--- A geometry that already has a CRS is returned unchanged
-SELECT ST_ApplyDefaultCRS(ST_SetCRS(ST_Point(-122.35, 47.65), 'EPSG:3857'));
-```
