+++ rust/sedona-functions/src/register.rs
@@ -82,7 +82,6 @@ pub fn default_function_set() -> FunctionSet {
         crate::st_rotate::st_rotate_y_udf,
         crate::st_scale::st_scale_udf,
         crate::st_setsrid::st_set_crs_udf,
-        crate::st_setsrid::st_apply_default_crs_udf,
         crate::st_setsrid::st_set_srid_udf,
         crate::st_srid::st_crs_udf,
         crate::st_srid::st_srid_udf,
