+++ rust/sedona-raster-functions/src/executor.rs
@@ -225,7 +225,7 @@ impl<'a, 'b> RasterExecutor<'a, 'b> {
     /// This is useful when the executor is built from a subset of the original
     /// arguments (e.g. only raster + geometry) but the overall UDF should still
     /// iterate according to other array arguments.
-    #[allow(dead_code)]
+    #[cfg(test)]
     pub fn new_with_num_iterations(
         arg_types: &'a [SedonaType],
         args: &'b [ColumnarValue],
